Calculate Quality-Adjusted Life Years

Discounted total QALYs up to a defined time horizon. This is a simpler, alternative function. See S3 Method also available (total_QALYs).

calc_QALY(utility = NA, intervals = NA, age = NA, start_delay = 0,
  discount_rate = 0.035, utility_method = "add")

Arguments

utility
Vector of values between 0 and 1 (1 - utility loss)

intervals
Time intervals for each utility, usually whole year i.e. 1. This is useful for fractions of years at the start and end of the period. However, since we may not know this then may not be necessary.

age
Year of age

start_delay
What time delay to origin, to shift discounting

discount_rate
Default: 3.5% per year

utility_method
How to combine utilities. Default: add, or prod

Details

Uses the following formula, for year i:

$$ \sum interval(i) * utility(i) * QoL(age(i)) * discount(i) $$

for i = 1, ..., time_horizon.

References

Sassi, Franco, Health Policy and Planning, 5, 402-408, Calculating QALYs, comparing QALY and DALY calculations, volume 21, 2006

See also

Examples

calc_QALY(utility = 0.9,
          age = 13,
          intervals = 49)

#> [[1]]
#>  [1] 0.8000000 0.7729469 0.7468086 0.7035153 0.6797249 0.6567391 0.6345305
#>  [8] 0.6052130 0.5847469 0.5649728 0.5458675 0.5274082 0.5095731 0.4923412
#> [15] 0.4756920 0.4596058 0.4440636 0.4290469 0.4145381 0.4005199 0.3869757
#> [22] 0.3738896 0.3612460 0.3490299 0.3372270 0.3258232 0.3148050 0.2962592
#> [29] 0.2862408 0.2765611 0.2672088 0.2581728 0.2494423 0.2410070 0.2328570
#> [36] 0.2249826 0.2173745 0.1960221 0.1893934 0.1829888 0.1768007 0.1708220
#> [43] 0.1650454 0.1594641 0.1540716 0.1488615 0.1438275 0.1389638 0.1342645
#>