Calculate annual inflation adjusted costs

Up to the present time inflated upwards.

inflation_adjust_cost(from_year, to_year, from_cost, reference = NA,
  inflation_data = 0.035)

Arguments

from_year
Date of cost to convert from

to_year
Date to convert cost to

from_cost
Cost at from_year

reference
Source of data (string)

inflation_data
NA default is fixed 3.5% rate of inflation, GDP_deflators, HCHS_pay or HCHS_price.

Value

Inflated cost (scalar), with attributes:

from_year

to_year

from_cost

reference

Details

Option to use the

https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/562750/GDP_Deflators_Qtrly_National_Accounts_September_2016_update_v2.csvONS GDP_Deflators_Qtrly_National_Accounts (This document contains the latest gross domestic product (GDP) deflators. The GDP deflator can be viewed as a measure of general inflation in the domestic economy)

https://www.pssru.ac.uk/pub/uc/uc2017/sources-of-information.pdfPSSRU annual inflation hospital and community health services

Fixed 3.5%.

Can't download directly into function because the .csv on the website is too messy as-is. This would be good to do though so that can always use latest version. ##TODO: webscraping? regular expressions?

$$(1 + i_{1})(1 + i_{2}) \cdots (1 + i_{n}) \times C$$

Examples

from_year <- 2012
to_year <- 2015
from_cost <- 96.140

inflation_adjust_cost(from_year,
                      to_year,
                      from_cost,
                      fixed = FALSE)

#> Error in inflation_adjust_cost(from_year, to_year, from_cost, fixed = FALSE): unused argument (fixed = FALSE)

#100

inflation_adjust_cost(from_year = 2010,
                      to_year = 2016,
                      from_cost = 1)

#> [1] 1.229255
#> attr(,"from_year")
#> [1] 2010
#> attr(,"to_year")
#> [1] 2016
#> attr(,"from_cost")
#> [1] 1
#> attr(,"reference")
#> [1] NA

#1.229255
1*(1+0.035)^6

#> [1] 1.229255