ONS National life tables: UK

Period life expectancy by age and sex for the UK. Each national life table is based on population estimates, births and deaths for a period of three consecutive years. Tables are published annually.

lifetable

Format

A data frame with 101 rows and 11 columns:

age

in years.

mx_male

Central rate of mortality, defined as the number of deaths at age \(x\) last birthday in the three year period to which the National Life Table relates divided by the average population at that age over the same period.

qx_male

Mortality rate between age \(x\) and \((x + 1)\), that is the probability that a person aged \(x\) exact will die before reaching age \((x + 1)\).

lx_male

Number of survivors to exact age \(x\) of 100,000 live births of the same sex who are assumed to be subject throughout their lives to the mortality rates experienced in the three year period to which the National Life Table relates.

dx_male

Number dying between exact age \(x\) and \((x + 1)\) described similarly to \(lx\), that is \(dx = lx - lx + 1\).

ex_male

Average period expectation of life at exact age \(x\), that is the average number of years that those aged \(x\) exact will live thereafter based on the mortality rates experienced in the three year period to which the National Life Table relates.

mx_female

Central rate of mortality, defined as the number of deaths at age \(x\) last birthday in the three year period to which the National Life Table relates divided by the average population at that age over the same period.

qx_female

Mortality rate between age \(x\) and \((x + 1)\), that is the probability that a person aged \(x\) exact will die before reaching age \((x + 1)\).

lx_female

Number of survivors to exact age \(x\) of 100,000 live births of the same sex who are assumed to be subject throughout their lives to the mortality rates experienced in the three year period to which the National Life Table relates.

dx_female

Number dying between exact age \(x\) and \((x + 1)\) described similarly to \(lx\), that is \(dx = lx - lx + 1\).

ex_female

Average period expectation of life at exact age \(x\), that is the average number of years that those aged \(x\) exact will live thereafter based on the mortality rates experienced in the three year period to which the National Life Table relates.

Source

https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/lifeexpectancies/datasets/nationallifetablesunitedkingdomreferencetables