Person health years object constructor

For use as input to the QALY (and DALY) functions. Want to provide set-up info in different ways but which gives us all the necessary information to go calc QALYs later. individual level data. Don’t need to supply all of these otherwise there would be redundancy and possible conflicts. E.g. time horizon can be got from start and end year. Need to calculate QoL which is discounting due to age from start to end year.

person_health_years(start_year = 0, end_year = NA, delay = 0,
  age = NA, time_horizon = NA, utility, discount_rate = 0.035,
  utility_method = "add")

Arguments

start_year
Calendar year to begin calculation

end_year
Calendar year to end calculation

delay
If the start_year is in the future then the first discounting is not 1 but smaller accordingly (positive integer) We could alternatively define year_now

age
Age at start of period (could be an age group)

time_horizon
Number of time periods from start to end date

utility
Proportion health detriment. This may change over time. Could provide as a list of length end-start or as say list(c(2,3), c(u1,u2)) like in WinBUGs data.

discount_rate
Fixed proportion reduction over time or a (continuous) function e.g. negative exponential

utility_method
Should the yearly QALYs be summed to a scalar? add or prod

Value

Object of class person_health_years

start_year

end_year

delay

age

time_horizon

utility

QoL

discount_rate

death

utility_method

See also

Examples

personHealthYears <- person_health_years(
                      start_year = 2016,
                      end_year = 2020,
                      age = 33,
                      time_horizon = NA,
                      utility = 0.9,
                      discount_rate = 0.035)

total_QALYs(personHealthYears)

#>   [1] 0.7700000 0.7439614 0.7188032 0.6944959        NA        NA        NA
#>   [8]        NA        NA        NA        NA        NA        NA        NA
#>  [15]        NA        NA        NA        NA        NA        NA        NA
#>  [22]        NA        NA        NA        NA        NA        NA        NA
#>  [29]        NA        NA        NA        NA        NA        NA        NA
#>  [36]        NA        NA        NA        NA        NA        NA        NA
#>  [43]        NA        NA        NA        NA        NA        NA        NA
#>  [50]        NA        NA        NA        NA        NA        NA        NA
#>  [57]        NA        NA        NA        NA        NA        NA        NA
#>  [64]        NA        NA        NA        NA        NA        NA        NA
#>  [71]        NA        NA        NA        NA        NA        NA        NA
#>  [78]        NA        NA        NA        NA        NA        NA        NA
#>  [85]        NA        NA        NA        NA        NA        NA        NA
#>  [92]        NA        NA        NA        NA        NA        NA        NA
#>  [99]        NA        NA
#> attr(,"class")
#> [1] "HRQoL"   "numeric"
#> attr(,"person_health_years")
#> $start_year
#> [1] 2016
#> 
#> $end_year
#> [1] 2020
#> 
#> $delay
#> [1] 0
#> 
#> $age
#> [1] 33
#> 
#> $time_horizon
#> [1] 4
#> 
#> $utility
#> [1] 0.9 0.9 0.9 0.9
#> 
#> $QoL
#> [1] 0.87 0.87 0.87 0.87 0.87
#> 
#> $discount_rate
#> [1] 0.035
#> 
#> $death
#> [1] NA
#> 
#> $utility_method
#> [1] "add"
#> 
#> $period
#> [1] 1 1 1 1
#> 
#> attr(,"class")
#> [1] "person_health_years" "list"               

## 2.913622

if (FALSE) {
 total_QALY(1)
 ## "Error: Not an person_health_years class input object."
}